Move ./cmd/server/internal contents into ./internal.
